Созданы: Класс для хранения данных объявления и Класс парсер который может собирать ссылки на объявления, открывать их и доставать нужную информацию.
